agent/pyproject.toml
[project]
name = "sample-multimodal-training-assistant-agent"
version = "0.1.0"
description = "Multimodal RAG agent with AG-UI and voice support"
authors = [{ name = "AG-UI Contributors" }]
requires-python = ">=3.12,<3.14"
dependencies = [
    "ag-ui-protocol>=0.1.5",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    "strands-agents[bidi]>=1.15.0",
    "bedrock-agentcore[strands-agents]>=0.1.21",
    "mcp-proxy-for-aws>=0.1.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
]

[tool.setuptools]
packages = ["agents", "lib", "tools"]
